feat: Create Test Deck with Test Card and spawn object
This commit creates a Test Deck with a Test Card and spawns the object in the game. The Test Deck is created using the `Forge.createDeck` function with the specified name, front image, back image, width, height, and cards. The `Forge.spawnObject` function is then used to spawn the object in the game at the specified position. This change improves the functionality of the application by adding the ability to create and spawn custom decks.